Moved fetcher back, since cleaner doesn't use it.
